Plugins (#254)
* Plugins

* CI fixes

* WIP

* Android fix

* Remove poll_events

* HPP Vulkan Triangle Fix

* Quality check fixes

* Android fix for hpp hello triangle rebase

* Add samples list

* Use ExitCodes instead of global try/catch

* Close records a close request if a window isnt yet available

* Android fix / better log message for premature close in android platform

* Fix android build

* Check for app bit requested outside of main loop

* todo: explain this

* change default window size

* change default window size

+++ samples/performance/msaa/msaa.cpp
@@ -746,7 +746,7 @@ void MSAASample::draw_gui()
 		    ImGui::PushItemWidth(ImGui::GetWindowWidth() * 0.4f);
 		    if (ImGui::BeginCombo("##sample_count", to_string(gui_sample_count).c_str()))
 		    {
-			    for (int n = 0; n < supported_sample_count_list.size(); n++)
+			    for (size_t n = 0; n < supported_sample_count_list.size(); n++)
 			    {
 				    bool is_selected = (gui_sample_count == supported_sample_count_list[n]);
 				    if (ImGui::Selectable(to_string(supported_sample_count_list[n]).c_str(), is_selected))
